refactor ts logic into flatter src directory
